Remove superflous multicall clients
