+++ packages/backend/src/modules/tvl/ArbitrumTvl.ts
@@ -9,9 +9,8 @@ import { BalanceUpdater } from '../../core/balances/BalanceUpdater'
 import { BlockNumberUpdater } from '../../core/BlockNumberUpdater'
 import { Clock } from '../../core/Clock'
 import { PriceUpdater } from '../../core/PriceUpdater'
-import { ArbitrumTotalSupplyProvider } from '../../core/totalSupply/providers/ArbitrumTotalSupplyProvider'
+import { TotalSupplyProvider } from '../../core/totalSupply/TotalSupplyProvider'
 import { TotalSupplyUpdater } from '../../core/totalSupply/TotalSupplyUpdater'
-import { ArbitrumMulticallClient } from '../../peripherals/arbitrum/multicall/ArbitrumMulticall'
 import { EthereumClient } from '../../peripherals/ethereum/EthereumClient'
 import { MulticallClient } from '../../peripherals/ethereum/MulticallClient'
 import { ARBITRUM_MULTICALL_CONFIG } from '../../peripherals/ethereum/MulticallConfig'
@@ -46,15 +45,14 @@ export function createArbitrumTvlSubmodule(
 
   const arbitrumClient = new EthereumClient(arbitrumProvider, logger, 25)
 
-  const arbitrumMulticall = ArbitrumMulticallClient.forMainnet(arbitrumClient)
   const multicallClient = new MulticallClient(
     arbitrumClient,
     ARBITRUM_MULTICALL_CONFIG,
   )
 
-  const totalSupplyProvider = new ArbitrumTotalSupplyProvider(
-    arbitrumClient,
-    arbitrumMulticall,
+  const totalSupplyProvider = new TotalSupplyProvider(
+    multicallClient,
+    ChainId.ARBITRUM,
   )
 
   const arbitrumBalanceProvider = new BalanceProvider(
